post covid-19

+++ objects/Object.php
@@ -1,36 +1,78 @@
 <?php
 
-interface ObjectInterface {
+interface ObjectInterface
+{
+    public static function getTableName();
 
-    static function getTableName();
-
-    static function getSearchFieldsNames();
+    public static function getSearchFieldsNames();
 }
 
 $tableExists = array();
 
-abstract class ObjectYPT implements ObjectInterface {
-
+abstract class ObjectYPT implements ObjectInterface
+{
     protected $fieldsName = array();
 
-    function __construct($id = "") {
+    public function __construct($id = "")
+    {
         if (!empty($id)) {
             // get data from id
             $this->load($id);
         }
     }
 
-    protected function load($id) {
+    protected function load($id)
+    {
         $row = self::getFromDb($id);
-        if (empty($row))
+        if (empty($row)) {
             return false;
+        }
         foreach ($row as $key => $value) {
             $this->$key = $value;
         }
         return true;
     }
 
-    static protected function getFromDb($id) {
+    public static function getNowFromDB()
+    {
+        global $global;
+        $sql = "SELECT NOW() as my_date_field";
+        $res = sqlDAL::readSql($sql);
+        $data = sqlDAL::fetchAssoc($res);
+        sqlDAL::close($res);
+        if ($res) {
+            $row = $data;
+        } else {
+            $row = false;
+        }
+        return $row;
+    }
+
+    public static function setTimeZone()
+    {
+        global $advancedCustom;
+        $row = self::getNowFromDB();
+        $dt = new DateTime($row['my_date_field']);
+        $timeZOnesOptions = object_to_array($advancedCustom->timeZone->type);
+        if (empty($timeZOnesOptions[$advancedCustom->timeZone->value])) {
+            return false;
+        }
+        try {
+            $objDate = new DateTimeZone($timeZOnesOptions[$advancedCustom->timeZone->value]);
+            if (is_object($objDate)) {
+                $dt->setTimezone($objDate);
+                date_default_timezone_set($timeZOnesOptions[$advancedCustom->timeZone->value]);
+                return $dt;
+            }
+            return false;
+        } catch (Exception $exc) {
+            _error_log("setTimeZone: ".$exc->getMessage(), AVideoLog::$ERROR);
+            return false;
+        }
+    }
+
+    protected static function getFromDb($id)
+    {
         global $global;
         $id = intval($id);
         $sql = "SELECT * FROM " . static::getTableName() . " WHERE  id = ? LIMIT 1";
@@ -46,7 +88,7 @@ abstract class ObjectYPT implements ObjectInterface {
         return $row;
     }
 
-    static function getAll() {
+    public static function getAll(){
         global $global;
         if (!static::isTableInstalled()) {
             return false;
@@ -67,8 +109,31 @@ abstract class ObjectYPT implements ObjectInterface {
         }
         return $rows;
     }
+    
+    public static function getAllActive(){
+        global $global;
+        if (!static::isTableInstalled()) {
+            return false;
+        }
+        $sql = "SELECT * FROM  " . static::getTableName() . " WHERE status='a' ";
 
-    static function getTotal() {
+        $sql .= self::getSqlFromPost();
+        $res = sqlDAL::readSql($sql);
+        $fullData = sqlDAL::fetchAllAssoc($res);
+        sqlDAL::close($res);
+        $rows = array();
+        if ($res != false) {
+            foreach ($fullData as $row) {
+                $rows[] = $row;
+            }
+        } else {
+            die($sql . '\nError : (' . $global['mysqli']->errno . ') ' . $global['mysqli']->error);
+        }
+        return $rows;
+    }
+
+    public static function getTotal()
+    {
         //will receive
         //current=1&rowCount=10&sort[sender]=asc&searchPhrase=
         global $global;
@@ -83,7 +148,8 @@ abstract class ObjectYPT implements ObjectInterface {
         return $countRow;
     }
 
-    static function getSqlFromPost($keyPrefix = "") {
+    public static function getSqlFromPost($keyPrefix = "")
+    {
         global $global;
         $sql = self::getSqlSearchFromPost();
 
@@ -93,7 +159,7 @@ abstract class ObjectYPT implements ObjectInterface {
             $_POST['sort'][$_GET['columns'][$index]['data']] = $_GET['order'][0]['dir'];
         }
 
-        // add a security here 
+        // add a security here
         if (!empty($_POST['sort'])) {
             foreach ($_POST['sort'] as $key => $value) {
                 $_POST['sort'][xss_esc($key)] = xss_esc($value);
@@ -119,7 +185,8 @@ abstract class ObjectYPT implements ObjectInterface {
         return $sql;
     }
 
-    static function getSqlLimit() {
+    public static function getSqlLimit()
+    {
         global $global;
         $sql = "";
 
@@ -129,7 +196,7 @@ abstract class ObjectYPT implements ObjectInterface {
 
         if (empty($_POST['current']) && !empty($_GET['start'])) {
             $_POST['current'] = ($_GET['start'] / $_GET['length']) + 1;
-        } else if (empty($_POST['current']) && isset($_GET['start'])) {
+        } elseif (empty($_POST['current']) && isset($_GET['start'])) {
             $_POST['current'] = 1;
         }
 
@@ -150,11 +217,12 @@ abstract class ObjectYPT implements ObjectInterface {
         return $sql;
     }
 
-    static function getSqlSearchFromPost() {
+    public static function getSqlSearchFromPost()
+    {
         $sql = "";
         if (!empty($_POST['searchPhrase'])) {
             $_GET['q'] = $_POST['searchPhrase'];
-        } else if (!empty($_GET['search']['value'])) {
+        } elseif (!empty($_GET['search']['value'])) {
             $_GET['q'] = $_GET['search']['value'];
         }
         if (!empty($_GET['q'])) {
@@ -178,7 +246,8 @@ abstract class ObjectYPT implements ObjectInterface {
         return $sql;
     }
 
-    function save() {
+    public function save()
+    {
         if (!$this->tableExists()) {
             _error_log("Save error, table " . static::getTableName() . " does not exists", AVideoLog::$ERROR);
             return false;
@@ -193,9 +262,9 @@ abstract class ObjectYPT implements ObjectInterface {
                     // do nothing
                 } elseif (strtolower($value) == 'modified') {
                     $fields[] = " {$value} = now() ";
-                } else if (is_numeric($this->$value)) {
+                } elseif (is_numeric($this->$value)) {
                     $fields[] = " `{$value}` = {$this->$value} ";
-                } else if (strtolower($this->$value) == 'null') {
+                } elseif (strtolower($this->$value) == 'null') {
                     $fields[] = " `{$value}` = NULL ";
                 } else {
                     $fields[] = " `{$value}` = '{$this->$value}' ";
@@ -234,7 +303,8 @@ abstract class ObjectYPT implements ObjectInterface {
         }
     }
 
-    private function getAllFields() {
+    private function getAllFields()
+    {
         global $global, $mysqlDatabase;
         $sql = "SELECT COLUMN_NAME FROM INFORMATION_SCHEMA.COLUMNS WHERE TABLE_SCHEMA = ? AND TABLE_NAME = '" . static::getTableName() . "'";
         $res = sqlDAL::readSql($sql, "s", array($mysqlDatabase));
@@ -251,7 +321,8 @@ abstract class ObjectYPT implements ObjectInterface {
         return $rows;
     }
 
-    function delete() {
+    public function delete()
+    {
         global $global;
         if (!empty($this->id)) {
             $sql = "DELETE FROM " . static::getTableName() . " ";
@@ -264,7 +335,8 @@ abstract class ObjectYPT implements ObjectInterface {
         return false;
     }
 
-    static function setCache($name, $value) {
+    public static function setCache($name, $value)
+    {
         $cachefile = self::getCacheFileName($name);
         make_path($cachefile);
         $bytes = @file_put_contents($cachefile, json_encode($value));
@@ -272,76 +344,121 @@ abstract class ObjectYPT implements ObjectInterface {
         return $bytes;
     }
 
-    static function cleanCacheName($name) {
-        $name = str_replace(array('/','\\'), array(DIRECTORY_SEPARATOR, DIRECTORY_SEPARATOR), $name);
-        $name = preg_replace('/[!#$&\'()*+,:;=?@[\\]% ]+/', '_', trim(strtolower(cleanString($name))));
+    public static function cleanCacheName($name)
+    {
+        $name = str_replace(array('/', '\\'), array(DIRECTORY_SEPARATOR, DIRECTORY_SEPARATOR), $name);
+        $name = preg_replace('/[!#$&\'()*+,:;=?@[\\]% -]+/', '_', trim(strtolower(cleanString($name))));
         $name = preg_replace('/\/{2,}/', '/', trim(strtolower(cleanString($name))));
         return preg_replace('/[\x00-\x1F\x7F]/u', '', $name);
     }
 
     /**
-     * 
+     *
      * @param type $name
      * @param type $lifetime, if is = 0 it is unlimited
      * @return type
      */
-    static function getCache($name, $lifetime = 60) {
+    public static function getCache($name, $lifetime = 60, $ignoreSessionCache=false)
+    {
+        if (isCommandLineInterface()) {
+            return false;
+        }
+        global $getCachesProcessed, $_getCache;
+
+        if (empty($_getCache)) {
+            $_getCache = array();
+        }
+
+        if (empty($getCachesProcessed)) {
+            $getCachesProcessed=array();
+        }
         $cachefile = self::getCacheFileName($name);
+
+        if (!empty($_getCache[$name])) {
+            return $_getCache[$name];
+        }
+
+        if (empty($getCachesProcessed[$name])) {
+            $getCachesProcessed[$name] = 0;
+        }
+        $getCachesProcessed[$name]++;
+
         if (!empty($_GET['lifetime'])) {
             $lifetime = intval($_GET['lifetime']);
         }
-        if(!empty($lifetime)){// do not session cache if there is not timeout limit
+
+        if (!empty($ignoreSessionCache)) {
             $session = self::getSessionCache($name, $lifetime);
             if (!empty($session)) {
+                $_getCache[$name] = $session;
                 return $session;
             }
         }
 
         if (file_exists($cachefile) && (empty($lifetime) || time() - $lifetime <= filemtime($cachefile))) {
             $c = @url_get_contents($cachefile);
-            return json_decode($c);
-        } else if (file_exists($cachefile)) {
+            $json = json_decode($c);
+            self::setSessionCache($name, $json);
+            $_getCache[$name] = $json;
+            return $json;
+        } elseif (file_exists($cachefile)) {
             self::deleteCache($name);
         }
     }
 
-    static function deleteCache($name) {
+    public static function deleteCache($name)
+    {
         $cachefile = self::getCacheFileName($name);
         @unlink($cachefile);
-
         self::deleteSessionCache($name);
+        ObjectYPT::deleteCacheFromPattern($name);
     }
 
-    static function deleteALLCache() {
+    public static function deleteALLCache()
+    {
         $tmpDir = self::getCacheDir();
         rrmdir($tmpDir);
         self::deleteAllSessionCache();
+        self::setLastDeleteALLCacheTime();
     }
 
-    static function getCacheDir() {
+    public static function getCacheDir()
+    {
         $tmpDir = getTmpDir();
-        $tmpDir = rtrim($tmpDir, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR;
-        $tmpDir .= "YPTObjectCache" . DIRECTORY_SEPARATOR;
-        
-        if(class_exists("User_Location")){
+        $tmpDir = rtrim($tmpDir, DIRECTORY_SEPARATOR) . "/";
+        $tmpDir .= "YPTObjectCache" . "/";
+
+        if (class_exists("User_Location")) {
             $loc = User_Location::getThisUserLocation();
-            if(!empty($loc)){
-                $tmpDir .= $loc['country_code'] . DIRECTORY_SEPARATOR;
+            if (!empty($loc) && !empty($loc['country_code'])) {
+                $tmpDir .= $loc['country_code'] . "/";
             }
         }
-        
+
         make_path($tmpDir);
-        if (!file_exists($tmpDir . "index.html")) {// to avoid search into the directory
+        if (!file_exists($tmpDir . "index.html") && is_writable($tmpDir)) {// to avoid search into the directory
             file_put_contents($tmpDir . "index.html", time());
         }
         return $tmpDir;
     }
 
-    static function getCacheFileName($name) {
+    public static function getCacheFileName($name)
+    {
         $name = self::cleanCacheName($name);
         $tmpDir = self::getCacheDir();
         $uniqueHash = md5(__FILE__);
         return $tmpDir . DIRECTORY_SEPARATOR . $name . $uniqueHash;
+    }
+
+    public static function deleteCacheFromPattern($name)
+    {
+        $name = self::cleanCacheName($name);
+        $tmpDir = self::getCacheDir();
+        $filePattern = $tmpDir . DIRECTORY_SEPARATOR . $name;
+        foreach (glob("{$filePattern}*") as $filename) {
+            unlink($filename);
+        }
+        self::deleteSessionCache($name);
     }
 
     /**
@@ -349,20 +466,25 @@ abstract class ObjectYPT implements ObjectInterface {
      * @param type $name
      * @param type $value
      */
-    static function setSessionCache($name, $value) {
+    public static function setSessionCache($name, $value)
+    {
         $name = self::cleanCacheName($name);
         _session_start();
         $_SESSION['user']['sessionCache'][$name]['value'] = json_encode($value);
         $_SESSION['user']['sessionCache'][$name]['time'] = time();
+        if (empty($_SESSION['user']['sessionCache']['time'])) {
+            $_SESSION['user']['sessionCache']['time'] = time();
+        }
     }
 
     /**
-     * 
+     *
      * @param type $name
      * @param type $lifetime, if is = 0 it is unlimited
      * @return type
      */
-    static function getSessionCache($name, $lifetime = 60) {
+    public static function getSessionCache($name, $lifetime = 60)
+    {
         $name = self::cleanCacheName($name);
         if (!empty($_GET['lifetime'])) {
             $lifetime = intval($_GET['lifetime']);
@@ -371,43 +493,98 @@ abstract class ObjectYPT implements ObjectInterface {
             if ((empty($lifetime) || time() - $lifetime <= $_SESSION['user']['sessionCache'][$name]['time'])) {
                 $c = $_SESSION['user']['sessionCache'][$name]['value'];
                 return json_decode($c);
-            } else {
-                _session_start();
-                unset($_SESSION['user']['sessionCache'][$name]);
             }
+            _session_start();
+            unset($_SESSION['user']['sessionCache'][$name]);
         }
+        return null;
     }
 
-    static function deleteSessionCache($name) {
+    public static function clearSessionCache()
+    {
+        unset($_SESSION['user']['sessionCache']);
+    }
+
+    private static function getLastDeleteALLCacheTimeFile()
+    {
+        $tmpDir = getTmpDir();
+        $tmpDir = rtrim($tmpDir, DIRECTORY_SEPARATOR) . "/";
+        $tmpDir .= "lastDeleteALLCacheTime.cache";
+        return $tmpDir;
+    }
+
+    public static function setLastDeleteALLCacheTime()
+    {
+        $file = self::getLastDeleteALLCacheTimeFile();
+        _error_log("ObjectYPT::setLastDeleteALLCacheTime {$file}");
+        return file_put_contents($file, time());
+    }
+
+    public static function getLastDeleteALLCacheTime()
+    {
+        global $getLastDeleteALLCacheTime;
+        if (empty($getLastDeleteALLCacheTime)) {
+            $getLastDeleteALLCacheTime = (int) @file_get_contents(self::getLastDeleteALLCacheTimeFile(), time());
+        }
+        return $getLastDeleteALLCacheTime;
+    }
+
+    public static function checkSessionCacheBasedOnLastDeleteALLCacheTime()
+    {
+        /*
+        var_dump(
+                $session_var['time'],
+                self::getLastDeleteALLCacheTime(),
+                humanTiming($session_var['time']),
+                humanTiming(self::getLastDeleteALLCacheTime()),
+                $session_var['time'] <= self::getLastDeleteALLCacheTime());
+         *
+         */
+        if (empty($_SESSION['user']['sessionCache']['time']) || $_SESSION['user']['sessionCache']['time'] <= self::getLastDeleteALLCacheTime()) {
+            self::deleteAllSessionCache();
+            return false;
+        }
+        return true;
+    }
+
+    public static function deleteSessionCache($name)
+    {
         $name = self::cleanCacheName($name);
         _session_start();
         $_SESSION['user']['sessionCache'][$name] = null;
         unset($_SESSION['user']['sessionCache'][$name]);
     }
 
-    static function deleteAllSessionCache() {
+    public static function deleteAllSessionCache()
+    {
         _session_start();
         unset($_SESSION['user']['sessionCache']);
     }
 
-    function tableExists() {
+    public function tableExists()
+    {
         return self::isTableInstalled();
     }
 
-    static function isTableInstalled($tableName = "") {
+    public static function isTableInstalled($tableName = "")
+    {
         global $global, $tableExists;
         if (empty($tableName)) {
             $tableName = static::getTableName();
         }
+        if (empty($tableName)) {
+            return false;
+        }
         if (!isset($tableExists[$tableName])) {
-            $res = sqlDAL::readSql("SHOW TABLES LIKE '" . $tableName . "'");
+            $sql = "SHOW TABLES LIKE '" . $tableName . "'";
+            //_error_log("isTableInstalled: ({$sql})");
+            $res = sqlDAL::readSql($sql);
             $result = sqlDal::num_rows($res);
             sqlDAL::close($res);
             $tableExists[$tableName] = !empty($result);
         }
         return $tableExists[$tableName];
     }
-
 }
 
 //abstract class Object extends ObjectYPT{};
